app/pyproject.toml
[project]
name = "inkdav"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "caldav==1.3.9",
    "icalendar==6.1.3",
    "Pillow==10.4.0",
    "Flask==3.0.3",
    "requests==2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths  = ["tests"]
